LayoutTests/accessibility/mac/search-predicate.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../../resources/accessibility-helper.js"></script>
</head>
<body>

<div id="content" role="group">

<blockquote>first block quote level 1</blockquote>
<blockquote>second block quote level 1</blockquote>
<p style="color:black; font-family:sans-serif; font-weight:bold; text-decoration:underline;">sans-serif black bold text with underline</p>
<p style="color:blue; font-family:serif; font-style:italic;">serif blue italic text</p>
<form>
<input type="text" /><br />
<input type="checkbox" value="Checkbox" /> checkbox<br />
<input type="submit" value="Submit" />
</form>
<img id="image" src="resources/cake.png" alt="Test">
<h1>first heading level 1</h1>
<h1>second heading level 1</h1>
<h2>heading level 2</h2>
<h3>heading level 3</h3>
<h4>heading level 4</h4>
<h5>heading level 5</h5>
<h6>heading level 6</h6>
<p id="banner" role="banner">landmark</p>
<a href="#image">link</a>
<ol><li>list item</li></ol>
<p aria-live="polite">live region</p>
<ul role="radiogroup"><li role="radio">radio button</li></ul>
<table>
<thead><tr><td>first table heading level 1</td></tr></thead>
<tbody><tr><td>first table body level 1</td></tr></tbody>
</table>
<table>
<thead><tr><td>second table heading level 1</td></tr></thead>
<tbody><tr><td>second table body level 1</td></tr></tbody>
</table>
<p>misspellleedd</p>
<ul role="tree"><li role="treeitem">tree item</li></ul>

<h1>isVisible test start</h1>
<div id="group" style="overflow: scroll; width:100px; height:100px;">
    <button id="button">test button 1</button><BR>
    test test test test <br>
    test test test test <br>
    test test test test <br>
    test test test test <br>
    test test test test <br>
    test test test test <br>
    test test test test <br>
    test test test test <br>
    test test test test <br>
    <button id="button">test button 2</button>
</div>
<button id="button">test button 3</button>

</div>

<script>
    let output = "This tests the ability to search for accessible elements by key or text.\n\n";

    if (window.accessibilityController) {
        window.jsTestIsAsync = true;

        window.containerElement = accessibilityController.accessibleElementById("content");
        // Any element.
        window.startElement = containerElement.childAtIndex(0);
        window.resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXAnyTypeSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXStaticText'");
        output += expect("resultElement.stringValue", "'AXValue: first block quote level 1'");

        // Same level blockquote.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXBlockquoteSameLevelSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXGroup'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: second block quote level 1'");

        // Blockquote.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXBlockquoteSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXGroup'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: second block quote level 1'");

        // Bold font.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXBoldFontSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXGroup'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: sans-serif black bold text with underline'");

        // Button.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXButtonSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXButton'");

        // Check box.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXCheckBoxSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXCheckBox'");

        // Control.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXCheckBoxSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXCheckBox'");

        // Different type element.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXDifferentTypeSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXStaticText'");
        output += expect("resultElement.stringValue", "'AXValue: first block quote level 1'");

        // Font change.
        startElement = containerElement.childAtIndex(2);
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXFontChangeSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXGroup'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: serif blue italic text'");

        // Font color change.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXFontColorChangeSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXGroup'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: serif blue italic text'");

        // Graphic.
        startElement = containerElement.childAtIndex(0);
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXGraphicSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXImage'");

        // Heading level 1.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXHeadingLevel1SearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXHeading'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: first heading level 1'");

        // Heading level 2.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXHeadingLevel2SearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXHeading'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: heading level 2'");

        // Heading level 3.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXHeadingLevel3SearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXHeading'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: heading level 3'");

        // Heading level 4.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXHeadingLevel4SearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXHeading'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: heading level 4'");

        // Heading level 5.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXHeadingLevel5SearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXHeading'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: heading level 5'");

        // Heading level 6.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXHeadingLevel6SearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXHeading'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: heading level 6'");

        // Same level heading.
        startElement = containerElement.childAtIndex(6);
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXHeadingSameLevelSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXHeading'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: second heading level 1'");

        // Heading.
        startElement = containerElement.childAtIndex(0);
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXHeadingSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXHeading'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: first heading level 1'");

        // Italic font.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXItalicFontSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXGroup'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: serif blue italic text'");

        // Landmark.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXLandmarkSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXGroup'");
        output += expect("resultElement.subrole", "'AXSubrole: AXLandmarkBanner'");

        // Link.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXLinkSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXLink'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: link'");

        // List.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXListSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXList'");

        // Live Region.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXLiveRegionSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXGroup'");
        output += expect("resultElement.stringAttributeValue('AXARIALive')", "'polite'");

        // Misspelled word.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXMisspelledWordSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXStaticText'");
        output += expect("resultElement.stringValue", "'AXValue: misspellleedd'");

        // Outline.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXOutlineSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXOutline'");
        output += expect("resultElement.childAtIndex(0).childAtIndex(0).stringValue", "'AXValue: tree item'");

        // Plain text.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXPlainTextSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXStaticText'");
        output += expect("resultElement.stringValue", "'AXValue: first block quote level 1'");

        // Radio group.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXRadioGroupSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXRadioGroup'");

        // Same type element.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXSameTypeSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXGroup'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: second block quote level 1'");

        // Static text.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXStaticTextSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXStaticText'");
        output += expect("resultElement.stringValue", "'AXValue: first block quote level 1'");

        // Style change.
        startElement = containerElement.childAtIndex(2);
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXStyleChangeSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXGroup'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: serif blue italic text'");

        // Same level table.
        startElement = containerElement.childAtIndex(18);
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXTableSameLevelSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXTable'");
        output += expect("resultElement.childAtIndex(0).childAtIndex(0).childAtIndex(0).stringValue", "'AXValue: second table heading level 1'");

        // Table.
        startElement = containerElement.childAtIndex(0);
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXTableSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXTable'");
        output += expect("resultElement.childAtIndex(0).childAtIndex(0).childAtIndex(0).stringValue", "'AXValue: first table heading level 1'");

        // Text field.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXTextFieldSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXTextField'");

        // Underline.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXUnderlineSearchKey", "", false);
        output += expect("resultElement.role", "'AXRole: AXGroup'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: sans-serif black bold text with underline'");

        // Unvisited link.
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXUnvisitedLinkSearchKey", "", false);
        output += expect("resultElement.boolAttributeValue('AXVisited')", "false");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: link'");

        // Previous text search.
        startElement = containerElement.childAtIndex(10);
        resultElement = containerElement.uiElementForSearchPredicate(startElement, false, "", "sans-serif black bold text with underline", false);
        output += expect("resultElement.role", "'AXRole: AXStaticText'");
        output += expect("resultElement.stringValue", "'AXValue: sans-serif black bold text with underline'");

        // Execute a search for the next heading level 2 or the next link.
        startElement = containerElement.childAtIndex(0);
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, ["AXHeadingLevel2SearchKey", "AXLinkSearchKey"], "", false);
        output += expect("resultElement.role", "'AXRole: AXHeading'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: heading level 2'");

        // After finding the heading, execute the search again and we should find the link.
        resultElement = containerElement.uiElementForSearchPredicate(resultElement, true, ["AXHeadingLevel2SearchKey", "AXLinkSearchKey"], "", false);
        output += expect("resultElement.role", "'AXRole: AXLink'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: link'");

        // From the link, execute the search in reverse and we should land back on the heading.
        resultElement = containerElement.uiElementForSearchPredicate(resultElement, false, ["AXHeadingLevel2SearchKey", "AXLinkSearchKey"], "", false);
        output += expect("resultElement.role", "'AXRole: AXHeading'");
        output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: heading level 2'");

        // Scroll all the way to the bottom of the content
        resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "", "test button 3", false);
        output += expect("resultElement.role", "'AXRole: AXButton'");
        output += expect("resultElement.title", "'AXTitle: test button 3'");
        setTimeout(async function() {
            resultElement.scrollToMakeVisible();
            await waitFor(() => {
                return resultElement.isOnScreen;
            });

            // find the start of the isVisible test section
            resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "", "isVisible test start", false);
            output += expect("resultElement.role", "'AXRole: AXHeading'");
            output += expect("resultElement.childAtIndex(0).stringValue", "'AXValue: isVisible test start'");

            // save away the "isVisible test start" heading as the start element
            startElement = resultElement;

            // If we don't care about visible only, then we should easily find 3 buttons
            resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXButtonSearchKey", "", false);
            output += expect("resultElement.role", "'AXRole: AXButton'");
            output += expect("resultElement.title", "'AXTitle: test button 1'");

            resultElement = containerElement.uiElementForSearchPredicate(resultElement, true, "AXButtonSearchKey", "", false);
            output += expect("resultElement.role", "'AXRole: AXButton'");
            output += expect("resultElement.title", "'AXTitle: test button 2'");

            // save away testButton2 so we can make it visible later
            window.testButton2 = resultElement;

            resultElement = containerElement.uiElementForSearchPredicate(resultElement, true, "AXButtonSearchKey", "", false);
            output += expect("resultElement.role", "'AXRole: AXButton'");
            output += expect("resultElement.title", "'AXTitle: test button 3'");

            // if we care about visible only, then we should not find "test button 2"
            resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXButtonSearchKey", "", true);
            output += expect("resultElement.role", "'AXRole: AXButton'");
            output += expect("resultElement.title", "'AXTitle: test button 1'");

            resultElement = containerElement.uiElementForSearchPredicate(resultElement, true, "AXButtonSearchKey", "", true);
            output += expect("resultElement.role", "'AXRole: AXButton'");
            output += expect("resultElement.title", "'AXTitle: test button 3'");

            // now, scroll to the second button, and confirm that we don't see the first button
            testButton2.scrollToMakeVisible();
            await waitFor(() => {
                return testButton2.isOnScreen;
            });

            resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXButtonSearchKey", "", true);
            output += expect("resultElement.role", "'AXRole: AXButton'");
            output += expect("resultElement.title", "'AXTitle: test button 2'");

            resultElement = containerElement.uiElementForSearchPredicate(resultElement, true, "AXButtonSearchKey", "", true);
            output += expect("resultElement.role", "'AXRole: AXButton'");
            output += expect("resultElement.title", "'AXTitle: test button 3'");

            // Now since the page is scrolled to the bottom, the first visible button should be #2
            startElement = containerElement.childAtIndex(0);
            resultElement = containerElement.uiElementForSearchPredicate(startElement, true, "AXButtonSearchKey", "", true);
            output += expect("resultElement.role", "'AXRole: AXButton'");
            output += expect("resultElement.title", "'AXTitle: test button 2'");

            debug(output);
            content.style.visibility = "hidden";
            finishJSTest();
        }, 0);
    }
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
